added deployment tasks, moved components into their respective folders and renamed their extensions to tsx
