+++ src/components/Header/header.tsx
@@ -2,7 +2,12 @@ import { Link } from 'gatsby'
 import PropTypes from 'prop-types'
 import React from 'react'
 
-const Header = ({ siteTitle }) => (
+interface Props {
+	/**Site Header Title */
+	siteTitle: string
+}
+
+const Header = ({ siteTitle }: Props) => (
 	<header
 		style={{
 			background: `rebeccapurple`,
