+++ src/components/Layout/layout.tsx
@@ -9,10 +9,15 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { useStaticQuery, graphql } from 'gatsby'
 
-import Header from './header'
+import Header from '../Header/header'
 import './layout.css'
 
-const Layout = ({ children }) => {
+interface Props {
+	/** Header Title */
+	children?: any
+}
+
+const Layout = ({ children }: Props) => {
 	const data = useStaticQuery(graphql`
 		query SiteTitleQuery {
 			site {
