+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/layout'
-import Image from '../components/image'
-import SEO from '../components/seo'
+import Layout from '../components/Layout/layout'
+import Image from '../components/Image/image'
+import SEO from '../components/Seo/seo'
 
 const IndexPage = () => (
 	<Layout>
