+++ src/pages/page-2.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/layout'
-import SEO from '../components/seo'
+import Layout from '../components/Layout/layout'
+import SEO from '../components/Seo/seo'
 
 const SecondPage = () => (
 	<Layout>
